updated integration event flow
